+++ client/web/gulpfile.js
@@ -12,9 +12,6 @@ const { createProxyMiddleware } = require('http-proxy-middleware')
 const signale = require('signale')
 const createWebpackCompiler = require('webpack')
 const WebpackDevServer = require('webpack-dev-server')
-// The `DevServerPlugin` should be exposed after the `webpack-dev-server@4` goes out of the beta stage.
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-const DevServerPlugin = require('webpack-dev-server/lib/utils/DevServerPlugin')
 
 const {
   graphQlSchema,
@@ -114,9 +111,10 @@ async function webpackDevelopmentServer() {
     port: DEV_SERVER_LISTEN_ADDR.port,
     // Disable default DevServer compression. We need more fine grained compression to support streaming search.
     compress: false,
-    onBeforeSetupMiddleware: developmentServer => {
+    setupMiddlewares: (middlewares, developmentServer) => {
       // Re-enable gzip compression using our own `compression` filter.
       developmentServer.app.use(compression({ filter: shouldCompressResponse }))
+      return middlewares
     },
     client: {
       overlay: false,
@@ -134,12 +132,6 @@ async function webpackDevelopmentServer() {
     },
     proxy: proxyConfig,
     webSocketServer: 'ws',
-  }
-
-  // Based on the update: https://github.com/webpack/webpack-dev-server/pull/2844
-  if (!webpackConfig.plugins.find(plugin => plugin.constructor === DevServerPlugin)) {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-    webpackConfig.plugins.push(new DevServerPlugin(options))
   }
 
   const compiler = createWebApplicationCompiler()