+++ internal/database/security_event_logs_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/sourcegraph/log/logtest"
 
+	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
 )
 
@@ -19,10 +20,10 @@ func TestSecurityEventLogs_ValidInfo(t *testing.T) {
 	t.Parallel()
 	logger, exportLogs := logtest.Captured(t)
 	db := NewDB(logger, dbtest.NewDB(logger, t))
-	ctx := context.Background()
 
 	var testCases = []struct {
 		name  string
+		actor *actor.Actor // optional
 		event *SecurityEvent
 		err   string
 	}{
@@ -32,9 +33,19 @@ func TestSecurityEventLogs_ValidInfo(t *testing.T) {
 			err:   `INSERT: ERROR: new row for relation "security_event_logs" violates check constraint "security_event_logs_check_name_not_empty" (SQLSTATE 23514)`,
 		},
 		{
-			name:  "InvalidUser",
-			event: &SecurityEvent{Name: "test_event", URL: "http://sourcegraph.com", Source: "WEB"},
-			err:   `INSERT: ERROR: new row for relation "security_event_logs" violates check constraint "security_event_logs_check_has_user" (SQLSTATE 23514)`,
+			name: "InvalidUser",
+			event: &SecurityEvent{Name: "test_event", URL: "http://sourcegraph.com", Source: "WEB",
+				// a UserID or AnonymousUserID is required to identify a user, unless internal
+				UserID: 0, AnonymousUserID: ""},
+			err: `INSERT: ERROR: new row for relation "security_event_logs" violates check constraint "security_event_logs_check_has_user" (SQLSTATE 23514)`,
+		},
+		{
+			name:  "InternalActor",
+			actor: &actor.Actor{Internal: true},
+			event: &SecurityEvent{Name: "test_event", URL: "http://sourcegraph.com", Source: "WEB",
+				// unset UserID and AnonymousUserID will error in other scenarios
+				UserID: 0, AnonymousUserID: ""},
+			err: "<nil>",
 		},
 		{
 			name:  "EmptySource",
@@ -64,6 +75,10 @@ func TestSecurityEventLogs_ValidInfo(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
+			ctx := context.Background()
+			if tc.actor != nil {
+				ctx = actor.WithActor(ctx, tc.actor)
+			}
 			err := db.SecurityEventLogs().Insert(ctx, tc.event)
 			got := fmt.Sprintf("%v", err)
 			assert.Equal(t, tc.err, got)
@@ -72,7 +87,7 @@ func TestSecurityEventLogs_ValidInfo(t *testing.T) {
 
 	logs := exportLogs()
 	auditLogs := filterAudit(logs)
-	assert.Equal(t, 3, len(auditLogs))
+	assert.Equal(t, 3, len(auditLogs)) // note: internal actor does not generate an audit log
 	for _, auditLog := range auditLogs {
 		assertAuditField(t, auditLog.Fields["audit"].(map[string]any))
 		assertEventField(t, auditLog.Fields["event"].(map[string]any))