+++ internal/repos/webhookworker/cleaner.go
@@ -23,13 +23,19 @@ func NewCleaner(ctx context.Context, observationCtx *observation.Context, worker
 		Metrics: metrics,
 	})
 
-	return goroutine.NewPeriodicGoroutineWithMetrics(ctx, 1*time.Hour, goroutine.NewHandlerWithErrorMessage(
-		"webhook_build_worker_cleaner",
-		func(ctx context.Context) error {
-			_, err := cleanJobs(ctx, workerBaseStore)
-			return err
-		},
-	), operation)
+	return goroutine.NewPeriodicGoroutineWithMetrics(
+		ctx,
+		"webhook.build_worker_cleaner",
+		"deletes completed and failed sync webhooks",
+		1*time.Hour,
+		goroutine.HandlerFunc(
+			func(ctx context.Context) error {
+				_, err := cleanJobs(ctx, workerBaseStore)
+				return err
+			},
+		),
+		operation,
+	)
 }
 
 func cleanJobs(ctx context.Context, workerBaseStore *basestore.Store) (numCleaned int, err error) {